+++ GoedeDoelenHelpen/Data/ApplicationDbContext.cs
@@ -1,8 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.ComponentModel.DataAnnotations;
-using System.Linq;
-using Microsoft.AspNetCore.Identity;
+﻿using System.Linq;
 using Microsoft.AspNetCore.Identity.EntityFrameworkCore;
 using Microsoft.EntityFrameworkCore;
 
@@ -71,220 +67,5 @@ namespace GoedeDoelenHelpen.Data
                 .WithMany(tp => tp.Donations)
                 .OnDelete(DeleteBehavior.Restrict);
         }
-    }
-
-    public class ApplicationUser: IdentityUser
-    {
-        public IEnumerable<Donation> Donations { get; set; }
-        
-        public IEnumerable<EventParticipant> EventParticipants { get; set; }
-
-        public IEnumerable<CharityApplicationUser> CharityApplicationUsers { get; set; }
-
-        public IEnumerable<TeamParticipantActivated> TeamParticipants { get; set; }
-    }
-
-    public class Charity
-    {
-        public int Id { get; set; }
-        [Required]
-        [StringLength(250)]
-        public string Adress { get; set; }
-        [Required]
-        [StringLength(250)]
-        public string Name { get; set; }
-        [Required]
-        [StringLength(maximumLength: 8, MinimumLength = 8)]
-        public string KVKNumber { get; set; }
-        /// <remarks>
-        /// check the official specs
-        /// </remarks>
-        [Required]        
-        public string FiscalNumber { get; set; }
-        /// <summary>
-        /// TODO: need format check
-        /// </summary>
-        [Required]
-        public string IBAN { get; set; }
-
-        /// <summary>
-        /// relation to intersection entity CharityApplicationUser
-        /// </summary>
-        public IEnumerable<CharityApplicationUser> CharityApplicationUsers { get; set; }
-
-        public IEnumerable<CharityDonation> Donations { get; set; }
-
-        public IEnumerable<CharityEvent> Events { get; set; }
-    }
-
-    public class CharityApplicationUser
-    {
-        public int Id { get; set; }
-
-        [Required]
-        public string ApplicationUserId { get; set; }
-        public ApplicationUser ApplicationUser { get; set; }
-
-        public int CharityId { get; set; }
-        public Charity Charity { get; set; }
-
-        public CharityApplicationUserRole CharityApplicationUserRole { get; set; }
-    }
-
-    public enum CharityApplicationUserRole
-    {
-        Admin,
-        User
-    }
-
-    public class Project
-    {
-        public int Id { get; set; }
-        [Required]
-        public string Name { get; set; }
-
-        public int CharityId { get; set; }
-        public Charity Charity { get; set; }
-
-        public IEnumerable<ProjectDonation> Donations { get; set; }
-
-        public IEnumerable<ProjectEvent> Events { get; set; }
-    }
-
-
-    /// <summary>
-    /// base type for all the different types of Donations
-    /// </summary>
-    /// <remarks>
-    /// EF will put all different donations in one table and uses a shadow property named the "discriminator" to determine wich type it actualy is.
-    /// </remarks>
-    public abstract class Donation
-    {
-        public int Id { get; set; }
-        [Range(0, 10_000)]
-        public decimal Amount { get; set; }
-        public DateTime DateTime { get; set; }
-
-        [Required]
-        public string ApplicationUserId { get; set; }
-        public ApplicationUser ApplicationUser { get; set; }
-        public string DonationType { get; internal set; }
-    }
-
-    public class CharityDonation: Donation
-    {
-        public int CharityId { get; set; }
-        public Charity Charity { get; set; }
-    }
-
-    public class ProjectDonation: Donation
-    {
-        public int ProjectId { get; set; }
-        public Project Project { get; set; }
-    }
-
-
-    public abstract class Event
-    {
-        public int Id { get; set; }
-        [Required]
-        public string Name { get; set; }
-
-        public IEnumerable<EventDonation> Donations { get; set; }
-    }
-
-    public class ProjectEvent: Event
-    {
-        public int ProjectId { get; set; }
-        public Project Project { get; set; }
-    }
-
-    public class CharityEvent: Event
-    {
-        public int CharityId { get; set; }
-        public Charity Charity { get; set; }
-    }
-
-    public class EventTeam
-    {
-        public int Id { get; set; }
-        [Required]
-        public string Name { get; set; }
-
-        public int EventId { get; set; }
-        public Event Event { get; set; }
-
-        public IEnumerable<TeamParticipant> TeamParticipants { get; set; }
-        public IEnumerable<EventTeamDonation> Donations { get; set; }
-    }
-
-    public enum TeamParticipantRole
-    {
-        member
-    }
-
-    public abstract class TeamParticipant
-    {
-        public int Id { get; set; }
-
-        public int EventId { get; set; }
-        public EventTeam EventTeam { get; set; }
-
-        public TeamParticipantRole TeamParticipantRole { get; set; }
-        public IEnumerable<TeamParticipantDonation> Donations { get; set; }
-    }
-
-    public class TeamParticipantActivated: TeamParticipant
-    {
-        [Required]
-        public string ApplicationUserId { get; set; }
-        public ApplicationUser ApplicationUser { get; set; }
-    }
-
-    public class TeamParticipantNotActivated: TeamParticipant
-    {
-        [Required]
-        [EmailAddress]
-        public string Email { get; set; }
-        [Required]
-        public string Name { get; set; }
-    }
-
-    public class EventParticipant
-    {
-        public int Id { get; set; }
-
-        public int EventId { get; set; }
-        public Event Event { get; set; }
-        
-        public string AplicationUserId { get; set; }
-+       public ApplicationUser ApplicationUser { get; set; }
-        
-        public IEnumerable<EventParticipantDonation> Donations { get; set; }
-    }
-
-
-    public class EventDonation: Donation
-    {
-        public int EventId { get; set; }
-        public Event Event { get; set; }
-    }
-
-    public class EventParticipantDonation: Donation
-    {
-        public int EventParticipantId { get; set; }
-        public EventParticipant EventParticipant { get; set; }
-    }
-
-    public class EventTeamDonation: Donation
-    {
-        public int EventTeamId { get; set; }
-        public EventTeam EventTeam { get; set; }
-    }
-
-    public class TeamParticipantDonation: Donation
-    {
-        public int TeamParticipantId { get; set; }
-        public TeamParticipant TeamParticipant { get; set; }
     }
 }
