+++ GoedeDoelenHelpen/Data/CharityApplicationUserRole.cs
@@ -1,0 +1,8 @@
+﻿namespace GoedeDoelenHelpen.Data
+{
+    public enum CharityApplicationUserRole
+    {
+        Admin,
+        User
+    }
+}
