+++ GoedeDoelenHelpen/Data/CharityDonation.cs
@@ -1,0 +1,8 @@
+﻿namespace GoedeDoelenHelpen.Data
+{
+    public class CharityDonation: Donation
+    {
+        public int CharityId { get; set; }
+        public Charity Charity { get; set; }
+    }
+}
