+++ GoedeDoelenHelpen/Data/CharityEvent.cs
@@ -1,0 +1,8 @@
+﻿namespace GoedeDoelenHelpen.Data
+{
+    public class CharityEvent: Event
+    {
+        public int CharityId { get; set; }
+        public Charity Charity { get; set; }
+    }
+}
