+++ GoedeDoelenHelpen/Data/EventDonation.cs
@@ -1,0 +1,8 @@
+﻿namespace GoedeDoelenHelpen.Data
+{
+    public class EventDonation: Donation
+    {
+        public int EventId { get; set; }
+        public Event Event { get; set; }
+    }
+}
