+++ GoedeDoelenHelpen/Data/EventParticipantDonation.cs
@@ -1,0 +1,8 @@
+﻿namespace GoedeDoelenHelpen.Data
+{
+    public class EventParticipantDonation: Donation
+    {
+        public int EventParticipantId { get; set; }
+        public EventParticipant EventParticipant { get; set; }
+    }
+}
