+++ GoedeDoelenHelpen/Data/EventTeamDonation.cs
@@ -1,0 +1,8 @@
+﻿namespace GoedeDoelenHelpen.Data
+{
+    public class EventTeamDonation: Donation
+    {
+        public int EventTeamId { get; set; }
+        public EventTeam EventTeam { get; set; }
+    }
+}
