+++ GoedeDoelenHelpen/Data/ProjectDonation.cs
@@ -1,0 +1,8 @@
+﻿namespace GoedeDoelenHelpen.Data
+{
+    public class ProjectDonation: Donation
+    {
+        public int ProjectId { get; set; }
+        public Project Project { get; set; }
+    }
+}
