+++ GoedeDoelenHelpen/Data/ProjectEvent.cs
@@ -1,0 +1,8 @@
+﻿namespace GoedeDoelenHelpen.Data
+{
+    public class ProjectEvent: Event
+    {
+        public int ProjectId { get; set; }
+        public Project Project { get; set; }
+    }
+}
