+++ GoedeDoelenHelpen/Data/TeamParticipantActivated.cs
@@ -1,0 +1,11 @@
+﻿using System.ComponentModel.DataAnnotations;
+
+namespace GoedeDoelenHelpen.Data
+{
+    public class TeamParticipantActivated: TeamParticipant
+    {
+        [Required]
+        public string ApplicationUserId { get; set; }
+        public ApplicationUser ApplicationUser { get; set; }
+    }
+}
