+++ GoedeDoelenHelpen/Data/TeamParticipantDonation.cs
@@ -1,0 +1,8 @@
+﻿namespace GoedeDoelenHelpen.Data
+{
+    public class TeamParticipantDonation: Donation
+    {
+        public int TeamParticipantId { get; set; }
+        public TeamParticipant TeamParticipant { get; set; }
+    }
+}
