+++ GoedeDoelenHelpen/Data/TeamParticipantRole.cs
@@ -1,0 +1,7 @@
+﻿namespace GoedeDoelenHelpen.Data
+{
+    public enum TeamParticipantRole
+    {
+        member
+    }
+}
